refactor: notification log state moved to LogsTable component
